Move Janus event types into api.go
It will then get picked up by the easyjson generator to speed up parsing.
